+++ pkg/ctl/context/delete_context_test.go
@@ -27,6 +27,11 @@ import (
 )
 
 func TestDeleteContextCmd(t *testing.T) {
+	if !IsCI() {
+		t.Skip("skip this test, it will modify the your local file")
+		return
+	}
+
 	home := utils.HomeDir()
 	path := fmt.Sprintf("%s/.config/pulsar/config", home)
 	defer os.Remove(path)
